+++ ciexyz/color.go
@@ -2,8 +2,8 @@ package ciexyz
 
 // Color represents a linear normalised colour in CIE XYZ space.
 type Color struct {
-	X float64
-	Y float64
-	Z float64
-	A float64
+	X float32
+	Y float32
+	Z float32
+	A float32
 }
